feat: extract all login functions
