+++ src/pages/Near.tsx
@@ -1,36 +1,18 @@
-import React, { Fragment, useCallback, useEffect, useState } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { randomBytes } from 'crypto';
-import { providers } from 'near-api-js';
 import type { AccountView } from 'near-api-js/lib/providers/provider';
-import {
-  verifyFullKeyBelongsToUser,
-  verifySignature,
-  type SignedMessage,
-  type SignMessageParams,
-} from '@near-wallet-selector/core';
 import styled from 'styled-components';
-
 import { useWalletSelector } from '../context/WalletSelectorContext';
 import Loading from '../components/common/Loading';
-import { getOrCreateKeypair } from '../auth/ed25519';
-import { ResponseData } from '../api/response';
-import apiClient from '../api';
-import {
-  LoginRequest,
-  LoginResponse,
-  NearSignatureMessageMetadata,
-  NodeChallenge,
-  Payload,
-  SignatureMessage,
-  SignatureMessageMetadata,
-  WalletMetadata,
-  WalletSignatureData,
-  WalletType,
-} from '../api/dataSource/NodeDataSource';
-import { setStorageNodeAuthorized } from '../auth/storage';
 import LoginIcon from '../components/common/LoginIcon';
 import translations from '../constants/en.global.json';
+import {
+  handleSignMessage,
+  handleSignOut,
+  handleSwitchAccount,
+  handleSwitchWallet,
+} from '../auth/nearLogin';
+import { useVerifyNear } from '../hooks/loginNear';
 
 export interface Message {
   premium: boolean;
@@ -178,37 +160,18 @@ const Wrapper = styled.div`
 `;
 
 export default function NearLogin() {
+  const t = translations.loginPage.nearLogin;
   const navigate = useNavigate();
   const { selector, accounts, modal, accountId } = useWalletSelector();
+  const { getAccount, verifyMessageBrowserWallet } = useVerifyNear({ accountId, selector });
   const [account, setAccount] = useState<Account | null>(null);
   const [loading, setLoading] = useState<boolean>(false);
   const [errorMessage, setErrorMessage] = useState('');
-  const t = translations.loginPage.nearLogin;
   const appName = 'me';
-
-  const getAccount = useCallback(async (): Promise<Account | null> => {
-    if (!accountId) {
-      return null;
-    }
-
-    const { network } = selector.options;
-    const provider = new providers.JsonRpcProvider({ url: network.nodeUrl });
-
-    return provider
-      .query<AccountView>({
-        request_type: 'view_account',
-        finality: 'final',
-        account_id: accountId,
-      })
-      .then((data: any) => ({
-        ...data,
-        account_id: accountId,
-      }));
-  }, [accountId, selector]);
 
   useEffect(() => {
     const timeoutId = setTimeout(() => {
-      verifyMessageBrowserWallet();
+      verifyMessageBrowserWallet(setErrorMessage);
     }, 500);
 
     return () => {
@@ -221,247 +184,12 @@ export default function NearLogin() {
     if (!accountId) {
       return setAccount(null);
     }
-
     setLoading(true);
-
     getAccount().then((nextAccount: any) => {
       setAccount(nextAccount);
       setLoading(false);
     });
   }, [accountId, getAccount]);
-
-  async function handleSignOut() {
-    if (!account) {
-      return;
-    }
-    const wallet = await selector.wallet();
-
-    wallet
-      .signOut()
-      .then(() => {
-        setAccount(null);
-      })
-      .catch((err: any) => {
-        console.log('Failed to sign out');
-        setErrorMessage('Failed to sign out');
-        console.error(err);
-      });
-  }
-
-  function handleSwitchWallet() {
-    modal.show();
-  }
-
-  function handleSwitchAccount() {
-    const currentIndex = accounts.findIndex((x) => x.accountId === accountId);
-    const nextIndex = currentIndex < accounts.length - 1 ? currentIndex + 1 : 0;
-
-    const nextAccountId = accounts[nextIndex]?.accountId;
-
-    selector.setActiveAccount(nextAccountId ?? '');
-
-    alert('Switched account to ' + nextAccountId);
-  }
-
-  const verifyMessage = useCallback(
-    async (
-      message: SignMessageParams,
-      signedMessage: SignedMessage,
-    ): Promise<boolean> => {
-      console.log('verifyMessage', { message, signedMessage });
-      try {
-        const verifiedSignature = verifySignature({
-          publicKey: signedMessage.publicKey,
-          signature: signedMessage.signature,
-          message: message.message,
-          nonce: message.nonce,
-          recipient: message.recipient,
-          callbackUrl: message.callbackUrl ?? '',
-        });
-        const verifiedFullKeyBelongsToUser = await verifyFullKeyBelongsToUser({
-          publicKey: signedMessage.publicKey,
-          accountId: signedMessage.accountId,
-          network: selector.options.network,
-        });
-
-        const isMessageVerified =
-          verifiedFullKeyBelongsToUser && verifiedSignature;
-
-        const resultMessage = isMessageVerified
-          ? 'Successfully verified'
-          : 'Failed to verify';
-
-        console.log(
-          `${resultMessage} signed message: '${
-            message.message
-          }': \n ${JSON.stringify(signedMessage)}`,
-        );
-
-        return isMessageVerified;
-      } catch (error) {
-        console.error('Error while verifying message', error);
-        setErrorMessage('Error while verifying message');
-        return false;
-      }
-    },
-    [selector.options.network],
-  );
-
-  const verifyMessageBrowserWallet = useCallback(async () => {
-    const urlParams = new URLSearchParams(window.location.hash.substring(1));
-    const accId = urlParams.get('accountId') as string;
-    const publicKey = urlParams.get('publicKey') as string;
-    const signature = urlParams.get('signature') as string;
-
-    if (!accId && !publicKey && !signature) {
-      console.error('Missing params in url.');
-      return;
-    }
-
-    const message: SignMessageParams = JSON.parse(
-      localStorage.getItem('message')!,
-    );
-
-    const state: SignatureMessageMetadata = JSON.parse(message.state!);
-
-    if (!state.publicKey) {
-      state.publicKey = publicKey;
-    }
-
-    const stateMessage: SignatureMessageMetadata = JSON.parse(state.message);
-    if (!stateMessage.publicKey) {
-      stateMessage.publicKey = publicKey;
-      state.message = JSON.stringify(stateMessage);
-    }
-
-    const signedMessage = {
-      accountId: accId,
-      publicKey,
-      signature,
-    };
-
-    const isMessageVerified: boolean = await verifyMessage(
-      message,
-      signedMessage,
-    );
-
-    const url = new URL(window.location.href);
-    url.hash = '';
-    url.search = '';
-    window.history.replaceState({}, document.title, url);
-    localStorage.removeItem('message');
-    if (isMessageVerified) {
-      const signatureMetadata: NearSignatureMessageMetadata = {
-        recipient: message.recipient,
-        callbackUrl: message.callbackUrl!,
-        nonce: message.nonce.toString('base64'),
-      };
-      const payload: Payload = {
-        message: state,
-        metadata: signatureMetadata,
-      };
-      const walletSignatureData: WalletSignatureData = {
-        payload: payload,
-        publicKey: publicKey,
-      };
-      const walletMetadata: WalletMetadata = {
-        wallet: WalletType.NEAR({
-          networkId: selector.options.network.networkId,
-        }),
-        signingKey: publicKey,
-      };
-      const loginRequest: LoginRequest = {
-        walletSignature: signature,
-        payload: walletSignatureData.payload!,
-        walletMetadata: walletMetadata,
-      };
-
-      await apiClient
-        .node()
-        .login(loginRequest)
-        .then((result: ResponseData<LoginResponse>) => {
-          console.log('result', result);
-          if (result.error) {
-            console.error('login error', result.error);
-            setErrorMessage(`Error while login: ${result.error}`);
-          } else {
-            setStorageNodeAuthorized();
-            navigate('/identity');
-            console.log('login success');
-          }
-        })
-        .catch(() => {
-          console.error('error while login');
-          setErrorMessage(`Error while login`);
-        });
-    } else {
-      console.error('Message not verified');
-      setErrorMessage('Message not verified');
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [verifyMessage]);
-
-  async function handleSignMessage() {
-    try {
-      const challengeResponseData: ResponseData<NodeChallenge> = await apiClient
-        .node()
-        .requestChallenge();
-      const { publicKey } = await getOrCreateKeypair();
-
-      if (challengeResponseData.error) {
-        console.log('requestChallenge api error', challengeResponseData.error);
-        return;
-      }
-
-      const wallet = await selector.wallet('my-near-wallet');
-
-      const challengeNonce =
-        challengeResponseData?.data?.nonce ?? randomBytes(32).toString('hex');
-
-      const nonce: Buffer = Buffer.from(challengeNonce, 'base64');
-      const recipient = appName;
-      const callbackUrl = window.location.href;
-      const challengeContextId = challengeResponseData.data?.contextId ?? null;
-      const nodeSignature = challengeResponseData.data?.nodeSignature ?? '';
-      const timestamp =
-        challengeResponseData.data?.timestamp ?? new Date().getTime();
-
-      const signatureMessage: SignatureMessage = {
-        nodeSignature,
-        publicKey: publicKey,
-      };
-      const message: string = JSON.stringify(signatureMessage);
-
-      const state: SignatureMessageMetadata = {
-        publicKey: publicKey,
-        nodeSignature,
-        nonce: nonce.toString('base64'),
-        contextId: challengeContextId,
-        timestamp,
-        message,
-      };
-
-      if (wallet.type === 'browser') {
-        console.log('browser');
-
-        localStorage.setItem(
-          'message',
-          JSON.stringify({
-            message,
-            nonce: [...nonce],
-            recipient,
-            callbackUrl,
-            state: JSON.stringify(state),
-          }),
-        );
-      }
-
-      await wallet.signMessage({ message, nonce, recipient, callbackUrl });
-    } catch (error) {
-      console.error('Error while signing message', error);
-      setErrorMessage('Error while signing message.');
-    }
-  }
 
   if (loading) {
     return <Loading />;
@@ -485,7 +213,17 @@ export default function NearLogin() {
                 <span className="account-id-value">{accountId}</span>
               </div>
             </div>
-            <div className="logout-button" onClick={handleSignOut}>
+            <div
+              className="logout-button"
+              onClick={() =>
+                handleSignOut({
+                  account,
+                  selector,
+                  setAccount,
+                  setErrorMessage,
+                })
+              }
+            >
               {t.logoutButtonText}
             </div>
           </div>
@@ -494,14 +232,22 @@ export default function NearLogin() {
         <div
           className={`padding-wrapper ${account ? 'account-wrapper-padding-lg' : 'account-wrapper-padding-sm'}`}
         >
-          <button className="action-button" onClick={handleSwitchWallet}>
+          <button
+            className="action-button"
+            onClick={() => handleSwitchWallet(modal)}
+          >
             {t.switchWalletButtonText}
           </button>
-          <button className="action-button" onClick={handleSignMessage}>
+          <button
+            className="action-button"
+            onClick={() =>
+              handleSignMessage({ selector, appName, setErrorMessage })
+            }
+          >
             {t.authButtonText}
           </button>
           {accounts.length > 1 && (
-            <button className="action-button" onClick={handleSwitchAccount}>
+            <button className="action-button" onClick={() => handleSwitchAccount({accounts, accountId, selector})}>
               {t.switchAccountButtonText}
             </button>
           )}
